Convert ParserBlockingJavaScript to an HtmlCheck
... and extract ParserBlockingScriptTag as a LiquidCheck
