+++ test/checks/parser_blocking_javascript_test.rb
@@ -58,18 +58,6 @@ class ParserBlockingJavaScriptTest < Minitest::Test
     END
   end
 
-  def test_script_tag_filter
-    offenses = analyze_theme(
-      ThemeCheck::ParserBlockingJavaScript.new,
-      "templates/index.liquid" => <<~END,
-        {{ 'foo.js' | asset_url | script_tag }}
-      END
-    )
-    assert_offenses(<<~END, offenses)
-      The script_tag filter is parser-blocking. Use a script tag with the async or defer attribute for better performance at templates/index.liquid:1
-    END
-  end
-
   def test_parser_blocking_script_over_multiple_lines
     offenses = analyze_theme(
       ThemeCheck::ParserBlockingJavaScript.new,
